chore: move send reset link module to worker
